Changed the initialization of gamma_pi and gamma_theta. These are now informed.
NOTE: need to figure out if the sampling L of should be roubust to all
possible hyper parameters values or not. Also I don't particulary like
the gamma_theta = 1 initialization. I seem to recall on Blei's lecture
that the parameters for dirichelet should be < 0.

Bug fix: the sample_Lj was previously changed to modify the label by
itselve. It was necessary to change sample_L just to call sampl_Lj on
all j labels instead of waiting for a result.